flair_model/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bpemb==0.3.3
certifi==2022.6.15
charset-normalizer==2.1.0
cloudpickle==2.1.0
conllu==4.5.2
cycler==0.11.0
Deprecated==1.2.13
Django==4.1
djangorestframework==3.13.1
filelock==3.7.1
flair==0.11.3
fonttools==4.34.4
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
huggingface-hub==0.8.1
hyperopt==0.2.7
idna==3.3
importlib-metadata==4.12.0
Janome==0.4.2
joblib==1.1.0
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lxml==4.9.1
Markdown==3.4.1
matplotlib==3.5.2
more-itertools==8.13.0
mpld3==0.3
networkx==2.8.5
numpy==1.23.1
overrides==3.1.0
packaging==21.3
Pillow==9.2.0
pptree==3.1
py4j==0.10.9.5
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
scikit-learn==1.1.1
scipy==1.9.0
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
smart-open==6.0.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
sqlparse==0.4.2
tabulate==0.8.10
threadpoolctl==3.1.0
tokenizers==0.12.1                           
asgiref==3.5.2
beautifulsoup4==4.11.1
bpemb==0.3.3
certifi==2022.6.15
charset-normalizer==2.1.0
cloudpickle==2.1.0
conllu==4.5.2
cycler==0.11.0
Deprecated==1.2.13
Django==4.1